refactor(cas): split cache models and errors
